Better security, more grammar and less problems

+++ API/v1/controllers/user/delete.js
@@ -41,6 +41,9 @@ exports.deleteUser = (req, res, next)=>{
     }
     userModel.findOne({_id: pathId}).exec()
         .then(result =>{
+            if(result.roles=="superadmin"){
+                return res.status(500).json({message: "Excluded from this function"})
+            }
             if(result && result.roles!="superadmin"){
                 return userModel.deleteOne({_id: pathId}).exec()
                     .then(removeResult=>{
